# api/manager/swagger.yaml
basePath: /
definitions:
...
  d7y_io_dragonfly_v2_manager_models.Config:
    properties:
      bio:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.User'
      user_id:
        type: integer
      value:
        type: string
    type: object
  d7y_io_dragonfly_v2_manager_models.JSONMap:
    additionalProperties: {}
    type: object
  d7y_io_dragonfly_v2_manager_models.Job:
    properties:
      args:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      bio:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      result:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      scheduler_clusters:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SchedulerCluster'
        type: array
      seed_peer_clusters:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeerCluster'
        type: array
      state:
        type: string
      task_id:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.User'
      user_id:
        type: integer
    type: object
...
  d7y_io_dragonfly_v2_manager_models.Peer:
    properties:
      build_platform:
        type: string
      created_at:
        type: string
      download_port:
        type: integer
      git_commit:
        type: string
      git_version:
        type: string
      host_name:
        type: string
      id:
        type: integer
      idc:
        type: string
      ip:
        type: string
      is_del:
        type: integer
      kernel_version:
        type: string
      location:
        type: string
      object_storage_port:
        type: integer
      os:
        type: string
      platform:
        type: string
      platform_family:
        type: string
      platform_version:
        type: string
      port:
        type: integer
      proxy_port:
        type: integer
      scheduler_cluster:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SchedulerCluster'
      scheduler_cluster_id:
        type: integer
      state:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  d7y_io_dragonfly_v2_manager_models.Scheduler:
    properties:
      config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      created_at:
        type: string
      features:
        items:
          type: string
        type: array
      host_name:
        type: string
      id:
        type: integer
      idc:
        type: string
      ip:
        type: string
      is_del:
        type: integer
      last_keep_alive_at:
        type: string
      location:
        type: string
      port:
        type: integer
      scheduler_cluster:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SchedulerCluster'
      scheduler_cluster_id:
        type: integer
      state:
        type: string
      updated_at:
        type: string
    type: object
  d7y_io_dragonfly_v2_manager_models.SchedulerCluster:
    properties:
      bio:
        type: string
      client_config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      created_at:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      is_del:
        type: integer
      jobs:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Job'
        type: array
      name:
        type: string
      peers:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Peer'
        type: array
      schedulers:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Scheduler'
        type: array
      scopes:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      seed_peer_clusters:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeerCluster'
        type: array
      updated_at:
        type: string
    type: object
  d7y_io_dragonfly_v2_manager_models.SeedPeer:
    properties:
      created_at:
        type: string
      download_port:
        type: integer
      host_name:
        type: string
      id:
        type: integer
      idc:
        type: string
      ip:
        type: string
      is_del:
        type: integer
      location:
        type: string
      object_storage_port:
        type: integer
      port:
        type: integer
      seed_peer_cluster:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeerCluster'
      seed_peer_cluster_id:
        type: integer
      state:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  d7y_io_dragonfly_v2_manager_models.SeedPeerCluster:
    properties:
      bio:
        type: string
      config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.JSONMap'
      created_at:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      jobs:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Job'
        type: array
      name:
        type: string
      scheduler_clusters:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SchedulerCluster'
        type: array
      seed_peer:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeer'
        type: array
      updated_at:
        type: string
    type: object
  d7y_io_dragonfly_v2_manager_models.User:
    properties:
      avatar:
        type: string
      bio:
        type: string
      configs:
        items:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.Config'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_del:
        type: integer
      location:
        type: string
      name:
        type: string
      phone:
        type: string
      state:
        type: string
      updated_at:
        type: string
    type: object
...
  d7y_io_dragonfly_v2_manager_types.CreateSeedPeerClusterRequest:
    properties:
      bio:
        type: string
      config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SeedPeerClusterConfig'
      name:
        type: string
    required:
    - config
    - name
    type: object
  d7y_io_dragonfly_v2_manager_types.CreateSeedPeerRequest:
    properties:
      download_port:
        type: integer
      host_name:
        type: string
      idc:
        type: string
      ip:
        type: string
      location:
        type: string
      object_storage_port:
        type: integer
      port:
        type: integer
      seed_peer_cluster_id:
        type: integer
      type:
        enum:
        - super
        type: string
    required:
    - download_port
    - host_name
    - ip
    - port
    - seed_peer_cluster_id
    - type
    type: object
...
  d7y_io_dragonfly_v2_manager_types.SeedPeerClusterConfig:
    properties:
      load_limit:
        maximum: 50000
        minimum: 1
        type: integer
    type: object
...
  d7y_io_dragonfly_v2_manager_types.UpdateSeedPeerClusterRequest:
    properties:
      bio:
        type: string
      config:
        $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.SeedPeerClusterConfig'
      name:
        type: string
    type: object
  d7y_io_dragonfly_v2_manager_types.UpdateSeedPeerRequest:
    properties:
      download_port:
        type: integer
      idc:
        type: string
      ip:
        type: string
      location:
        type: string
      object_storage_port:
        type: integer
      port:
        type: integer
      seed_peer_cluster_id:
        type: integer
      type:
        enum:
        - super
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    url: https://d7y.io
  description: Dragonfly Manager Server
  license:
    name: Apache 2.0
  title: Dragonfly Manager
  version: 1.0.0
paths:
...
  /api/v1/seed-peer-clusters:
    get:
      consumes:
      - application/json
      description: Get SeedPeerClusters
      parameters:
      - default: 0
        description: current page
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: return max item count, default 10, max 50
        in: query
        maximum: 50
        minimum: 2
        name: per_page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeerCluster'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get SeedPeerClusters
      tags:
      - SeedPeerCluster
    post:
      consumes:
      - application/json
      description: Create by json config
      parameters:
      - description: DNCluster
        in: body
        name: SeedPeerCluster
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.CreateSeedPeerClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeerCluster'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create SeedPeerCluster
      tags:
      - SeedPeerCluster
  /api/v1/seed-peer-clusters/{id}:
    delete:
      consumes:
      - application/json
      description: Destroy by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Destroy SeedPeerCluster
      tags:
      - SeedPeerCluster
    get:
      consumes:
      - application/json
      description: Get SeedPeerCluster by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeerCluster'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get SeedPeerCluster
      tags:
      - SeedPeerCluster
    patch:
      consumes:
      - application/json
      description: Update by json config
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: SeedPeerCluster
        in: body
        name: SeedPeerCluster
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.UpdateSeedPeerClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeerCluster'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update SeedPeerCluster
      tags:
      - SeedPeerCluster
  /api/v1/seed-peer-clusters/{id}/scheduler-clusters/{scheduler_cluster_id}:
    put:
      consumes:
      - application/json
      description: Add SchedulerCluster to SeedPeerCluster
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: scheduler cluster id
        in: path
        name: scheduler_cluster_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Add SchedulerCluster to SeedPeerCluster
      tags:
      - SeedPeerCluster
  /api/v1/seed-peer-clusters/{id}/seed-peers/{seed_peer_id}:
    put:
      consumes:
      - application/json
      description: Add SeedPeer to SeedPeerCluster
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: seed peer id
        in: path
        name: seed_peer_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Add Instance to SeedPeerCluster
      tags:
      - SeedPeerCluster
  /api/v1/seed-peers:
    get:
      consumes:
      - application/json
      description: Get SeedPeers
      parameters:
      - default: 0
        description: current page
        in: query
        name: page
        required: true
        type: integer
      - default: 10
        description: return max item count, default 10, max 50
        in: query
        maximum: 50
        minimum: 2
        name: per_page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeer'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get SeedPeers
      tags:
      - SeedPeer
    post:
      consumes:
      - application/json
      description: Create by json config
      parameters:
      - description: SeedPeer
        in: body
        name: SeedPeer
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.CreateSeedPeerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeer'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Create SeedPeer
      tags:
      - SeedPeer
  /api/v1/seed-peers/{id}:
    delete:
      consumes:
      - application/json
      description: Destroy by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Destroy SeedPeer
      tags:
      - SeedPeer
    get:
      consumes:
      - application/json
      description: Get SeedPeer by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeer'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get SeedPeer
      tags:
      - SeedPeer
    patch:
      consumes:
      - application/json
      description: Update by json config
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: SeedPeer
        in: body
        name: SeedPeer
        required: true
        schema:
          $ref: '#/definitions/d7y_io_dragonfly_v2_manager_types.UpdateSeedPeerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/d7y_io_dragonfly_v2_manager_models.SeedPeer'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update SeedPeer
      tags:
      - SeedPeer
...
swagger: "2.0"
...